put each test into its own .c file
