+++ tester/tester.c
@@ -4,11 +4,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "qmk_wrappers.h"
-#include "keybuffer.h"
-#include "key_stack.h"
-#include "trie.h"
 #include "utils.h"
-#include "sequence_transform.h"
 #include "sequence_transform_data.h"
 #include "sequence_transform_test.h"
 #include "tester_utils.h"
@@ -50,120 +46,6 @@ void tap_code16(uint16_t keycode)
         default:
             sim_output_push(st_keycode_to_char(keycode));
     }   
-}
-//////////////////////////////////////////////////////////////////
-// (partial) simulation of process_sequence_transform logic
-// to test st_perform
-void sim_st_perform(const uint16_t *keycodes)
-{
-    // we don't use st_key_buffer_reset(buf) here because
-    // we don't nec want a space at the start of the buffer
-    st_key_buffer_t *buf = st_get_key_buffer();
-    buf->context_len = 0;
-    sim_output_reset();
-    for (uint16_t key = *keycodes; key; key = *++keycodes) {
-        st_key_buffer_push(buf, key);
-        // If st_perform doesn't do anything special with this key,
-        // add it to our virtual output buffer
-        if (!st_perform()) {
-            tap_code16(key);
-        }
-    }
-}
-//////////////////////////////////////////////////////////////////
-// (partial) simulation of process_sequence_transform logic
-// to test enhanced backspace
-void sim_st_enhanced_backspace(const uint16_t *keycodes)
-{
-    for (uint16_t key = *keycodes; key; key = *++keycodes) {
-        tap_code16(KC_BSPC);
-        st_handle_backspace();
-    }
-}
-//////////////////////////////////////////////////////////////////////
-static char missed_rule_seq[SEQUENCE_MAX_LENGTH + 1] = {0};
-static char missed_rule_transform[TRANSFORM_MAX_LEN + 1] = {0};
-void sequence_transform_on_missed_rule_user(const st_trie_rule_t *rule)
-{
-    strcpy_s(missed_rule_seq, sizeof(missed_rule_seq), rule->sequence);
-    strcpy_s(missed_rule_transform, sizeof(missed_rule_transform), rule->transform);
-}
-//////////////////////////////////////////////////////////////////////
-// (partial) simulation of process_sequence_transform logic
-// to test st_find_missed_rule
-void sim_st_find_missed_rule(const uint16_t *keycodes)
-{
-    missed_rule_seq[0] = 0;
-    missed_rule_transform[0] = 0;
-    st_key_buffer_t *buf = st_get_key_buffer();
-    // reset search_len_from_space by first sending a single space
-    st_key_buffer_push(buf, KC_SPACE);
-    st_find_missed_rule();
-    // send input rule seq so we can get output transform to test
-    sim_st_perform(keycodes);
-    buf->context_len = 0;
-    // send the output into input buffer
-    // to simulate user typing it directly
-    char *output = sim_output_get(false);
-    for (char c = *output; c; c = *++output) {
-        const uint16_t key = ascii_to_keycode(&c);
-        st_key_buffer_push(buf, key);
-    }
-    // from this new input buffer, find missed rule
-    st_find_missed_rule();
-}
-//////////////////////////////////////////////////////////////////////
-void test_perform(const st_test_rule_t *rule, st_test_result_t *res)
-{
-    static char message[256];
-    res->message = message;
-    // Test #1: st_perform
-    sim_st_perform(rule->seq_keycodes);
-    // Ignore spaces at the start of output
-    char *output = sim_output_get(true);
-    // Check if our output buffer matches the expected transform string
-    res->pass = !strcmp(output, rule->transform_str);
-    if (res->pass) {
-        sprintf_s(message, sizeof(message), "OK!");
-    } else {
-        sprintf_s(message, sizeof(message), "output: %s", output);
-    }
-}
-//////////////////////////////////////////////////////////////////////
-void test_backspace(const st_test_rule_t *rule, st_test_result_t *res)
-{
-    static char message[256];
-    res->message = message;
-    // Test #2: st_handle_backspace
-    // Make sure enhanced backspace handling leaves us with an empty
-    // output buffer if we send one backspace for every key sent
-    sim_st_enhanced_backspace(rule->seq_keycodes);
-    const int out_size = sim_output_get_size();
-    res->pass = out_size == 0;
-    if (res->pass) {
-        sprintf_s(message, sizeof(message), "OK!");
-    } else {
-        sprintf_s(message, sizeof(message), "left %d keys in buffer!", out_size);
-    }
-}
-//////////////////////////////////////////////////////////////////////
-void test_find_rule(const st_test_rule_t *rule, st_test_result_t *res)
-{
-    static char message[256];
-    res->message = message;
-    // Test #3: st_find_missed_rule
-    sim_st_find_missed_rule(rule->seq_keycodes);
-    res->pass = !strcmp(missed_rule_transform, rule->transform_str);
-    if (res->pass) {
-        sprintf_s(message, sizeof(message), "OK!");
-    } else {
-        if (strlen(missed_rule_seq)) {
-            sprintf_s(message, sizeof(message), "found: %s ⇒ %s",
-                missed_rule_seq, missed_rule_transform);
-        } else {
-            sprintf_s(message, sizeof(message), "found nothing!");
-        }
-    }
 }
 //////////////////////////////////////////////////////////////////////
 int test_rule(const st_test_rule_t *rule, bool print_all)
