+++ tester/tester.vcxproj
@@ -164,6 +164,9 @@
     <ClCompile Include="sim_output_buffer.c" />
     <ClCompile Include="tester.c" />
     <ClCompile Include="tester_utils.c" />
+    <ClCompile Include="test_backspace.c" />
+    <ClCompile Include="test_find_rule.c" />
+    <ClCompile Include="test_perform.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\keybuffer.h" />
